Cleanup PR #16
